[rANS] Extend dataset metrics
Use dataset metrics to estimate size of the dictionary

+++ Utilities/rANS/include/rANS/definitions.h
@@ -28,8 +28,8 @@ using symbol_t = int32_t;
 using count_t = uint32_t;
 using histogram_t = std::vector<count_t>;
 
-inline constexpr uint8_t MinRenormThreshold = 10;
-inline constexpr uint8_t MaxRenormThreshold = 20;
+inline constexpr size_t MinRenormThreshold = 10;
+inline constexpr size_t MaxRenormThreshold = 20;
 } // namespace rans
 } // namespace o2
